move barcode into its own file/capability
