+++ packages/teams-js/src/public/audioVisualDevice.ts
@@ -1,12 +1,7 @@
 /* eslint-disable @typescript-eslint/explicit-member-accessibility */
 
 import { sendAndHandleSdkError, sendMessageToParent, sendMessageToParentAsync } from '../internal/communication';
-import {
-  getMediaCallbackSupportVersion,
-  mediaAPISupportVersion,
-  scanBarCodeAPIMobileSupportVersion,
-} from '../internal/constants';
-import { GlobalVars } from '../internal/globalVars';
+import { getMediaCallbackSupportVersion, mediaAPISupportVersion } from '../internal/constants';
 import { registerHandler, removeHandler } from '../internal/handlers';
 import { ensureInitialized, isCurrentSDKVersionAtLeast } from '../internal/internalAPIs';
 import {
@@ -14,7 +9,6 @@ import {
   decodeAttachment,
   throwExceptionIfMediaCallIsNotSupportedOnMobile,
   validateGetMediaInputs,
-  validateScanBarCodeInput,
   validateSelectMediaInputs,
   validateViewImagesInput,
 } from '../internal/mediaUtil';
@@ -23,7 +17,7 @@ import {
   generateGUID,
   InputFunction,
 } from '../internal/utils';
-import { FrameContexts, HostClientType, MediaType } from './constants';
+import { FrameContexts, MediaType } from './constants';
 import { ErrorCode, ImageProps, SdkError } from './interfaces';
 import { media } from './media';
 import { runtime } from './runtime';
@@ -256,61 +250,6 @@ export namespace audioVisualDevice {
 
     export function isSupported(): boolean {
       return runtime.supports.media.camera ? true : false;
-    }
-
-    export namespace barcode {
-      /**
-       * Scan Barcode/QRcode using camera
-       *
-       * @remarks
-       * Note: For desktop and web, this API is not supported. Callback will be resolved with ErrorCode.NotSupported.
-       *
-       * @param config - optional input configuration to customize the barcode scanning experience
-       * @returns A promise resolved with the barcode data or rejected with an @see SdkError
-       */
-      export function scanBarCode(config?: BarCodeConfig): Promise<string> {
-        ensureInitialized(FrameContexts.content, FrameContexts.task);
-
-        return new Promise<string>(resolve => {
-          if (
-            GlobalVars.hostClientType === HostClientType.desktop ||
-            GlobalVars.hostClientType === HostClientType.web ||
-            GlobalVars.hostClientType === HostClientType.rigel ||
-            GlobalVars.hostClientType === HostClientType.teamsRoomsWindows ||
-            GlobalVars.hostClientType === HostClientType.teamsRoomsAndroid ||
-            GlobalVars.hostClientType === HostClientType.teamsPhones ||
-            GlobalVars.hostClientType === HostClientType.teamsDisplays
-          ) {
-            throw { errorCode: ErrorCode.NOT_SUPPORTED_ON_PLATFORM };
-          }
-
-          if (!isCurrentSDKVersionAtLeast(scanBarCodeAPIMobileSupportVersion)) {
-            throw { errorCode: ErrorCode.OLD_PLATFORM };
-          }
-
-          if (!validateScanBarCodeInput(config)) {
-            throw { errorCode: ErrorCode.INVALID_ARGUMENTS };
-          }
-
-          resolve(sendAndHandleSdkError('media.scanBarCode', config));
-        });
-      }
-
-      /**
-       * Barcode configuration supplied to scanBarCode API to customize barcode scanning experience in mobile
-       * All properties in BarCodeConfig are optional and have default values in the platform
-       */
-      export interface BarCodeConfig {
-        /**
-         * Optional; Lets the developer specify the scan timeout interval in seconds
-         * Default value is 30 seconds and max allowed value is 60 seconds
-         */
-        timeOutIntervalInSec?: number;
-      }
-
-      export function isSupported(): boolean {
-        return runtime.supports.media.camera.barcode ? true : false;
-      }
     }
   }
 }
